# strings: use Builder in ToUpper and ToLower